Extracted WEBGL client into cell-viz
