+++ examples/webgl/index.html
@@ -13,7 +13,6 @@
         <script src="https://code.jquery.com/jquery-3.1.0.slim.min.js"></script>
         <script src="../../bin/cell-viz-1.1.0.bundle.js"></script>
         <script src="../../bin/highbrow-0.0.1.bundle.js"></script>
-        <script src="webgl-rendering-adapter.js"></script>
 
         <script>
 
@@ -39,7 +38,7 @@
                 var network = Highbrow.createHtmNetwork(simple)
                 var layer = network.getCorticalColumns()[0].getLayers()[0]
 
-                var cellviz = new SingleLayerVisualization(new HtmHighbrowLayer(layer));
+                var cellviz = new HighbrowLayerVisualization(layer);
 
                 animateCells(layer);
