+++ src/layer.js
@@ -1,13 +1,16 @@
 // Highbrow
 // MIT License (see LICENSE)
-// Copyright © 2005—2017 Numenta <http://numenta.com>
+// Copyright © 2017 Numenta <http://numenta.com>
 
 /** @ignore */
 const Renderable = require("./renderable")
 /** @ignore */
-const times = require("./utils").times
+const Neuron = require("./neuron")
 /** @ignore */
 const NeuronState = require("./enums").NeuronState
+
+/** @ignore */
+const times = require("./utils").times
 
 /*
  * Active cell indices returned from HTM systems generally are ordered with
@@ -19,13 +22,13 @@ const NeuronState = require("./enums").NeuronState
  * @param {integer} rx - range of the x dimension
  * @param {integer} ry - range of the y dimension
  * @param {integer} rz - range of the z dimension
- * @return {Object} point with 3D coordinates
- * @property {number} x x coordinate
- * @property {number} y y coordinate
- * @property {number} z z coordinate
+ * @return {Object} The position (not coordinate)
+ * @property {integer} x x position
+ * @property {integer} y y position
+ * @property {integer} z z position
  */
  /** @ignore */
-function getXyzFromIndex(idx, xsize, ysize) {
+function getXyzPositionFromIndex(idx, xsize, ysize) {
     var zcapacity = xsize * ysize;
     var x = 0, y = 0, z = 0;
     if (idx >= zcapacity) {
@@ -116,23 +119,26 @@ class Layer extends Renderable {
         return out
     }
 
-    /*
-     * Builds out the layer from scratch using the config object.
+    /**
+     * Builds out the layer from scratch using the config object. Creates an
+     * array of {@link Neuron}s that will be used for the lifespan of the Layer.
      */
     _buildLayer() {
         this._neurons = []
         let count = this._config.neuronCount
         let scale = this.getScale()
         for (let i = 0; i < count; i++) {
-            this._neurons.push(new Neuron({
+            let neuron = new Neuron({
                 name: `Neuron ${i}`,
                 state: NeuronState.inactive,
-                origin: getXyzFromIndex(i,
+                index: i,
+                position: getXyzPositionFromIndex(i,
                     this._config.dimensions.x,
                     this._config.dimensions.y
                 ),
                 scale: scale
-            }, this))
+            }, this)
+            this._neurons.push(neuron)
         }
         if (this._config.miniColumns) {
             // TODO: implement minicolumns.
@@ -148,57 +154,6 @@ class MiniColumn extends Renderable {
     constructor(config, parent) {
         super(config, parent)
     }
-}
-
-/**
- * Represents a pyramidal neuron. The atomic unit of HTM computation.
- */
-class Neuron extends Renderable {
-    constructor(config, parent) {
-        super(config, parent)
-        this._state = NeuronState.inactive
-    }
-
-    activate() {
-        this._state = NeuronState.active
-    }
-
-    deactivate() {
-        this._state = NeuronState.inactive
-    }
-
-    /**
-     * @override NOOP
-     * @returns [] empty list
-     */
-    getChildren() {
-        return []
-    }
-
-    /**
-     * @override
-     */
-    getName() {
-        return `${this.index} (${this.state})`
-    }
-
-    /**
-     * @override
-     */
-    toString() {
-        let n = this.getName()
-        let o = this.getOrigin()
-        let s = this.getScale()
-        return `${n} at [${o.x}, ${o.y}, ${o.z}] (scaled by ${s})`
-    }
-
-    set state (state)  { this._state = state }
-
-    get state () { return this._state }
-
-    // This index only changes if the config changes (unlikely).
-    get index () { return this.getConfig()["index"] }
-
 }
 
 module.exports = Layer
