+++ src/renderable.js
@@ -1,6 +1,6 @@
 // Highbrow
 // MIT License (see LICENSE)
-// Copyright © 2005—2017 Numenta <http://numenta.com>
+// Copyright © 2017 Numenta <http://numenta.com>
 
 /** @ignore */
 const DEFAULT_ORIGIN = {x: 0, y: 0, z: 0}
